Individual Filters

EqualizerFilters.PK — Function

PK(f, dbGain, Q, fs=sampling_rate)

Return peaking biquad with Gain dbGain in dB at frequency f in Hz with quality factor Q.

In this implementation the Gain at 0 Hz and at the Nyquist frequency will always be 0 dB. For an Orfanidis-like peak EQ with analog-inspired Nyquist-gain see the package LakeBiquads.jl.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.

EqualizerFilters.LS — Function

LS(f, dbGain, Q, fs=sampling_rate)

Return low shelf biquad with Gain dbGain in dB, center frequency f in Hz and quality factor Q.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.

EqualizerFilters.HS — Function

HS(f, dbGain, Q, fs=sampling_rate)

Return high shelf biquad with Gain dbGain in dB, center frequency f in Hz and quality factor Q.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.

EqualizerFilters.LP — Function

LP(f, Q, fs=sampling_rate)

Return lowpass biquad at frequency f in Hz with quality factor Q.

This second order lowpass falls of with 12 dB/oct.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.

EqualizerFilters.HP — Function

HP(f, Q, fs=sampling_rate)

Return highpass biquad at frequency f in Hz with quality factor Q.

This second order highpass falls of with 12 dB/oct.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.

EqualizerFilters.BP — Function

BP(f, Q, fs=sampling_rate)

Return bandpass biquad at frequency f in Hz with quality factor Q.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.

EqualizerFilters.NO — Function

NO(f, Q, fs=sampling_rate)

Return notch biquad at frequency f in Hz with quality factor Q.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.

EqualizerFilters.AP — Function

AP(f, Q, fs=sampling_rate)

Return allpass biquad at frequency f in Hz with quality factor Q.

This second order allpass rotates the phase by a total of $2π$.

If not given in the arguments, the sampling rate defaults to 48 kHz or can be set with EqualizerFilters.sampling_rate.