Sample Rate control

EqualizerFilters.sampling_rate — Function

EqualizerFilters.sampling_rate(custom_sampling_rate)

Set the sampling rate for calculation of the biquads in Hz.

Defaults to 48 kHz.

EqualizerFilters.sampling_rate()

Return the current sampling rate for biquad calculation in Hz.

Defaults to 48 kHz.

The functions are not exported by default, so you'll usually have to call them as EqualizerFilters.sampling_rate().